modify protocol version
Former-commit-id: f1e713f0ac9b7da703a666410deb47abe9b410ba

+++ src/main.cpp
@@ -2410,7 +2410,7 @@ bool ProcessBlock(CValidationState &state, CNode* pfrom, CBlock* pblock, CDiskBl
     {
         LogPrint("INFO","ProcessBlock: ORPHAN BLOCK %lu hash=%s, prev=%s\n", (unsigned long)mapOrphanBlocks.size(), pblock->GetHash().GetHex(), pblock->hashPrevBlock.ToString());
 
-        if (pblock->nHeight > g_nSyncTipHeight)
+        if (pblock->nHeight > (unsigned int)g_nSyncTipHeight)
 			g_nSyncTipHeight = pblock->nHeight;
         // Accept orphans as long as there is a node to request its parents from
         if (pfrom) {
